refactor(manifests): clarify subsystem boundaries
