+++ nemoguardrails/manifests/__init__.py
@@ -13,12 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Public API for the rail manifest contract and catalog.
-
-Re-exports the manifest types (`RailManifest`, `RailMetadata`,
-`RailSpec`, and friends) and the `RailCatalog`, and provides
-process-wide accessors for the default catalog of built-in rails.
-"""
+"""Public API for rail manifests, catalog access, and configured surface references."""
 
 from nemoguardrails.manifests.catalog import RailCatalog, RailManifestRecord
 from nemoguardrails.manifests.manifest import (
@@ -45,41 +40,16 @@ from nemoguardrails.manifests.manifest import (
     TransformTarget,
     import_ref_target,
     iter_manifest_import_refs,
-    normalize_configured_surface_name,
-    parse_configured_surface,
     resolve_import_ref,
 )
-
-_catalog: RailCatalog | None = None
-_discovering = False
-
-
-def default_rail_catalog() -> RailCatalog:
-    """Return the cached catalog of built-in rail manifests."""
-    global _catalog, _discovering
-    if _catalog is not None:
-        return _catalog
-    if _discovering:
-        raise RuntimeError("Built-in rail manifest discovery re-entered while loading rail modules.")
-    _discovering = True
-    try:
-        catalog = RailCatalog.discover_built_ins()
-        _catalog = catalog
-        return catalog
-    finally:
-        _discovering = False
-
-
-def all_rail_manifests():
-    """Return built-in rail manifests keyed by manifest name."""
-    return dict(default_rail_catalog().manifests)
-
-
-def _reset_rail_manifest_cache() -> None:
-    global _catalog, _discovering
-    _catalog = None
-    _discovering = False
-
+from nemoguardrails.manifests.registry import (
+    all_rail_manifests,
+    default_rail_catalog,
+)
+from nemoguardrails.manifests.surface_reference import (
+    normalize_configured_surface_name,
+    parse_configured_surface,
+)
 
 __all__ = [
     "ActionRef",
