+++ nemoguardrails/manifests/catalog.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""In-memory registry of rail manifests and the surfaces they expose.
+"""Catalog construction and discovery for rail manifests and their surfaces.
 
 Collects `RailManifestRecord` entries for built-in rails discovered under
 `nemoguardrails/library` into an immutable `RailCatalog` that enforces global
